LED Flashes properly
LED now flashes by reading the string of MACE characters, at a speed
based on the potentiometer.

TO DO: Modify the potentiometer min and max values (as per the
assignment brief).
Set LED to flash RED if the string was translated, or BLUE if the string
was input as MACE characters.