Detecting commands.
Now correctly detects input commands. LP will ONLY run if the first two
letters are LP, and there are exactly 12 digits following. All other
commands run as expected.

To do - find out what constitutes a high or low value for the Infrared
receiver and program this. Also need to program the LP command to alter
the PWM values of the LEDs

The end is nigh